// src/app/services/openapi-client/model/couponDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CouponDto { 
    id?: number;
    code?: string | null;
    type?: string | null;
    value?: number;
    minPurchaseAmount?: number | null;
    maxDiscountAmount?: number | null;
    usageLimitTotal?: number | null;
    usageLimitPerUser?: number | null;
    timesUsed?: number;
    validFrom?: string | null;
    validUntil?: string | null;
    isActive?: boolean;
}